# WebAPISamples/DefaultWebAPI/Models/Employee.cs
namespace DefaultWebAPI.Models
{
    public class Employee
    {
        public int Id { get; set; } 
        public string FirstName { get;set; }
        public string LastName { get; set; }
    }
}